Partial results hook: consistency of param. col.
Always print the parameter coloumn if there will be more than one
DGPP.

Otherwise, the parameter column is omitted in the first (few)
chunks, and then once the second parameter is being run the
parameter column is printed. It looks more consistent to just always
print it.

The extra argument could have been added to mc_table(), but
mc_table() already has a lot of arguments so it is added to the
print function.